Remove all non-base classes from 'Development' class category
The 'Development' category is a way of flagging classes in the base
package that are design time only and to be removed from deployed apps.
Any other development tools can be removed from deployed apps if they
are not in prerequisite packages anway. We don't really want to forcibly
remove these any more since it is permissible to ship the dev tools now.